Fix bridges build

+++ source/bridges/carla_bridge_ui-lv2.cpp
@@ -463,7 +463,7 @@ public:
     // ---------------------------------------------------------------------
     // processing
 
-    void setParameter(const int32_t rindex, const double value)
+    void setParameter(const int32_t rindex, const float value)
     {
         CARLA_ASSERT(handle && descriptor);
